+++ src/components/exercise/Thumbnail.jsx
